tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/multiple_animations.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - multiple animations for the same property</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animations">
   <meta name="flags" content="dom">
   <meta name="assert" content="The animation defined last overrides previously
   defined animations for the same property">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes grow {
         from {
            width: 200px;
         }
         to {
            width: 200px;
         }
      }
      @keyframes shrink {
         from {
            width: 50px;
         }
         to {
            width: 50px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
      }
      .test {
         animation: shrink 1s, grow 1s;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function(evt) {
            if (evt.animationName == 'grow') {
               clearTimeout(timeout);
               var divWidth = getComputedStyle(div, null).width;
               var passed = divWidth == '200px';
               test(function(){assert_true(passed);}, testName); done();
               log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
            }
         });
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue square below should become a wide rectangle for 1 second, then
   return to its square shape.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
